Refactor wandac code to own function in CLI
